convert 'hal_float' to a double-precision type.
as discussed at length, pentium-and-above systems can safely write 'double's
with the required atomicitiy guarantee, and 'gcc -Os' appears always to emit
safe instruction sequences (which are also higher performance on post-pentium
systems)

For more details,
http://linuxcnc.org/irc/irc.freenode.net:6667/emcdevel/2008-11-04.txt
http://axis.unpy.net/01225680565

+++ src/hal/utils/halcmd_commands.c
@@ -564,7 +564,7 @@ int do_newsig_cmd(char *name, char *type)
 static int set_common(hal_type_t type, void *d_ptr, char *value) {
     // This function assumes that the mutex is held
     int retval = 0;
-    float fval;
+    double fval;
     long lval;
     unsigned long ulval;
     char *cp = value;
